i40evf: retire I40EVF_FLAG_ADDR_SET_BY_PF flag
A trusted VF is allowed to override the administratively set MAC
address. However the VF does not know if it is being trusted or not.
Therefore the I40EVF_FLAG_ADDR_SET_BY_PF is no longer useful as it
currently prevents i40evf to change the MAC address if an
administratively set MAC is detected.

Signed-off-by: Stefan Assmann <sassmann@kpanic.de>